Update logic.txt
Created "related" section. Included wikipedia link to target the variety of meaning to trivium.

In this file, and in general the whitestone, is meant to display modernized definitions -- soon or eventually to be addressed in the top level readme.md -- as opposed to expound on their historic context which is what wikipedia is often useful for. Because, the history of a subject, at times, can be extraneous, as is reflected in many academic subject where history, outside of the subject of history itself, is treated very briefly, and left to the student to discover on their own outside of the classroom. Likewise, historical coverage and caveat will most often fall outside the scope of this project.